Fix a bug in Disk_t: ZeroMemory() does not work for non POD.
Disk_t is NOT a POD as it contains a std::string.
Simply zeroing the memory is not a replacement for a constructor or for a .clear() method.
This would cause access violation when the string strFilenameInZip is used (e.g. line 365).

It is possible that in VS a std::string full of zeros is valid.
This can be triggered inserting the same floppy in both drives.

Signed-off-by: Andrea Odetti <mariofutire@gmail.com>